# src/api/docs/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "TTHK Timetable API",
    "description": "This is the unofficial API to access TTHK's timetable data",
    "version": "1.0.0"
  },
  "host": "localhost:5637",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/timetable/{timetableId}/classes/": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getAllClasses",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "description": "The name of the class to look for",
            "required": false,
            "type": "string"
          },
          {
            "name": "color",
            "in": "query",
            "description": "The color of the class to look for",
            "required": false,
            "type": "string"
          },
          {
            "name": "teachers",
            "in": "query",
            "description": "An array of teacher ids that a class must have assigned to them, which is typically only one",
            "required": false,
            "type": "array"
          },
          {
            "name": "classrooms",
            "in": "query",
            "description": "An array of classroom ids that a class must have assigned to them",
            "required": false,
            "type": "array"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableClassArray"
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/timetable/{timetableId}/classes/{classId}": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getClassById",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "classId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableClass"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/timetable/{timetableId}/classes/{classId}/teachers": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getClassTeachers",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "classId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableTeacherArray"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/timetable/{timetableId}/classes/{classId}/classrooms": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getClassAssignedRooms",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "classId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableClassroomArray"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/timetable/{timetableId}/lessons/": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getAllLessons",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "subject",
            "in": "query",
            "description": "The id of the subject that the lesson should be associated with",
            "required": false,
            "type": "string"
          },
          {
            "name": "teachers",
            "in": "query",
            "description": "An array of teacher ids that the lesson must have assigned to them",
            "required": false,
            "type": "array"
          },
          {
            "name": "rooms",
            "in": "query",
            "description": "An array of classroom ids that the lesson must have assigned to them",
            "required": false,
            "type": "array"
          },
          {
            "name": "period_start",
            "in": "query",
            "description": "Which period the lesson has to start at",
            "required": false,
            "type": "integer"
          },
          {
            "name": "period_end",
            "in": "query",
            "description": "Which period the lesson has to end at",
            "required": false,
            "type": "integer"
          },
          {
            "name": "days",
            "in": "query",
            "description": "An array of the names of days that the lesson is on",
            "required": false,
            "type": "array"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableLessonArray"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/timetable/{timetableId}/lessons/{lessonId}": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getLessonById",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableLesson"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/timetable/{timetableId}/lessons/{lessonId}/teachers": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getLessonTeachers",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableTeacherArray"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/timetable/{timetableId}/lessons/{lessonId}/subject": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getLessonSubject",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableSubject"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/timetable/{timetableId}/lessons/{lessonId}/rooms": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getLessonRooms",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableClassroomArray"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/timetable/{timetableId}/lessons/{lessonId}/period_span": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getLessonPeriodSpan",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/timetable/{timetableId}/rooms/": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getAllClassrooms",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "description": "The name of the room to look for",
            "required": false,
            "type": "string"
          },
          {
            "name": "short_name",
            "in": "query",
            "description": "The short name of the room to look for",
            "required": false,
            "type": "string"
          },
          {
            "name": "buildingid",
            "in": "query",
            "description": "The building id of the room to look for",
            "required": false,
            "type": "string"
          },
          {
            "name": "color",
            "in": "query",
            "description": "The color of the room to look for, not sure why anyone would use this",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableClassroomArray"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/timetable/{timetableId}/rooms/{roomId}": {
      "get": {
        "tags": [
          "Timetable"
        ],
        "description": "",
        "operationId": "getClassroomById",
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimetableClassroom"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "TimetableClass": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": ""
        },
        "name": {
          "type": "string",
          "example": ""
        },
        "color": {
          "type": "string",
          "example": ""
        },
        "teachers": {
          "type": "array",
          "example": [
            ""
          ],
          "items": {
            "type": "string"
          }
        },
        "classrooms": {
          "type": "array",
          "example": [
            ""
          ],
          "items": {
            "type": "string"
          }
        }
      }
    },
    "TimetableClassArray": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": ""
          },
          "name": {
            "type": "string",
            "example": ""
          },
          "color": {
            "type": "string",
            "example": ""
          },
          "teachers": {
            "type": "array",
            "example": [
              ""
            ],
            "items": {
              "type": "string"
            }
          },
          "classrooms": {
            "type": "array",
            "example": [
              ""
            ],
            "items": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "TimetableTeacherArray": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": ""
          },
          "name": {
            "type": "string",
            "example": ""
          },
          "color": {
            "type": "string",
            "example": ""
          },
          "taught_classes": {
            "type": "array",
            "example": [
              ""
            ],
            "items": {
              "type": "string"
            }
          }
        }
      }
    },
    "TimetableClassroom": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": ""
        },
        "name": {
          "type": "string",
          "example": ""
        },
        "short_name": {
          "type": "string",
          "example": ""
        },
        "buildingid": {
          "type": "string",
          "example": ""
        },
        "color": {
          "type": "string",
          "example": ""
        }
      }
    },
    "TimetableClassroomArray": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": ""
          },
          "name": {
            "type": "string",
            "example": ""
          },
          "short_name": {
            "type": "string",
            "example": ""
          },
          "buildingid": {
            "type": "string",
            "example": ""
          },
          "color": {
            "type": "string",
            "example": ""
          }
        }
      }
    },
    "TimetableLesson": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": ""
        },
        "subject": {
          "type": "string",
          "example": ""
        },
        "teachers": {
          "type": "array",
          "example": [
            ""
          ],
          "items": {
            "type": "string"
          }
        },
        "rooms": {
          "type": "array",
          "example": [
            ""
          ],
          "items": {
            "type": "string"
          }
        },
        "period_start": {
          "type": "number",
          "example": 0
        },
        "period_end": {
          "type": "number",
          "example": 0
        },
        "days": {
          "type": "array",
          "example": [
            ""
          ],
          "items": {
            "type": "string"
          }
        }
      }
    },
    "TimetableLessonArray": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": ""
          },
          "subject": {
            "type": "string",
            "example": ""
          },
          "teachers": {
            "type": "array",
            "example": [
              ""
            ],
            "items": {
              "type": "string"
            }
          },
          "rooms": {
            "type": "array",
            "example": [
              ""
            ],
            "items": {
              "type": "string"
            }
          },
          "period_start": {
            "type": "number",
            "example": 0
          },
          "period_end": {
            "type": "number",
            "example": 0
          },
          "days": {
            "type": "array",
            "example": [
              ""
            ],
            "items": {
              "type": "string"
            }
          }
        }
      }
    },
    "TimetableSubject": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": ""
        },
        "name": {
          "type": "string",
          "example": ""
        },
        "short_name": {
          "type": "string",
          "example": ""
        },
        "color": {
          "type": "string",
          "example": ""
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}